refactor(server): fold fixedActionArgs into the REST route declaration module
`http/rest-tool-action.ts` held a single three-line function and existed for
the same reason `http/rest-tool-routes.ts` does — keeping fixed-action REST
proxying importable without pulling the tool graph into unit tests. One
concern, one module, one test file.
